requirements.txt
blinker==1.9.0cachetools==5.5.2certifi==2025.1.31charset-normalizer==3.4.1click==8.1.8colorama==0.4.6contourpy==1.3.1cycler==0.12.1Flask==3.1.0Flask-SQLAlchemy==3.1.1fonttools==4.56.0google-auth==2.38.0google-auth-oauthlib==1.2.1greenlet==3.1.1gspread==6.2.0gunicorn==20.1.0honcho==2.0.0httplib2==0.22.0idna==3.10itsdangerous==2.2.0Jinja2==3.1.6kiwisolver==1.4.8MarkupSafe==3.0.2matplotlib==3.10.1narwhals==1.31.0numpy==2.2.3oauth2client==4.1.3oauthlib==3.2.2packaging==24.2pandas==2.2.3Pillow==11.1.0plotly==6.0.1psycopg2-binary==2.9.10pyasn1==0.6.1pyasn1_modules==0.4.1pyodbc==5.2.0pyparsing==3.2.1python-dateutil==2.9.0.post0pytz==2025.1requests==2.32.3requests-oauthlib==2.0.0rsa==4.9six==1.117.0SQLAlchemy==2.0.39typing_extensions==4.12.2tzdata==2025.1urllib3==2.3.0Werkzeug==3.1.3dash==3.0.1